Start to build a usable framework for downloading / parsing
Use requests / BeautifulSoup to get the list of all catalog files in the
EDR3.  Use wget to download each file in turn, then numpy.genformtxt to
convert the CSV file into a Numpy NDARRAY for actual parsing.

Next steps will be to check for Declination and Magnitude for inclusion
in the LDT's catalog.