Change file downloader
Switched out requersts for wget in terms of downloading the Gaia catalog
files form the server.  Wget was hanging somewhat regularly, and using requests
in this fashion seems to be more fault-tolerant.  Uses tqdm to produce a progress bar,
and this iteration is, in fact, more informative since tqdm also displays the time
remaining and current download speed.

Add tqdm to the project requirements (and remove wget).