Add ability to retry download of Gaia catalog file
Add a while loop to check for whether a given Gaia catalog file downloaded
correctly.  If not, try again.  Add a timeout = 10s to the requests.get object
in for each file to be downloaded to see if the script will not hang, but rather
abort and retry.  Add try/except error handling to catch the timeout error thrown,
and be sure to reload the http stream again and restart the progress bar to
start over.